Renamed some views to make them clearer
