+++ src/views/Categories.vue
